CORE-32995 Broke out createExecuteCommand.
